Extract replicaset controller to own package
