+++ pkg/virt-controller/watch/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "application.go",
         "migration.go",
         "migrationpolicy.go",
-        "replicaset.go",
         "util.go",
         "vm.go",
         "vmi.go",
@@ -52,6 +51,7 @@ go_library(
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
         "//pkg/virt-controller/watch/node:go_default_library",
         "//pkg/virt-controller/watch/pool:go_default_library",
+        "//pkg/virt-controller/watch/replicaset:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//pkg/virt-controller/watch/workload-updater:go_default_library",
@@ -105,7 +105,6 @@ go_test(
     srcs = [
         "application_test.go",
         "migration_test.go",
-        "replicaset_test.go",
         "vm_test.go",
         "vmi_test.go",
         "vsock_test.go",
@@ -130,6 +129,7 @@ go_test(
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
         "//pkg/virt-controller/watch/node:go_default_library",
+        "//pkg/virt-controller/watch/replicaset:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
