+++ pkg/virt-controller/watch/application.go
@@ -38,6 +38,7 @@ import (
 	"kubevirt.io/kubevirt/pkg/virt-controller/watch/clone"
 	"kubevirt.io/kubevirt/pkg/virt-controller/watch/node"
 	"kubevirt.io/kubevirt/pkg/virt-controller/watch/pool"
+	"kubevirt.io/kubevirt/pkg/virt-controller/watch/replicaset"
 
 	"kubevirt.io/kubevirt/pkg/instancetype"
 
@@ -167,7 +168,7 @@ type VirtControllerApp struct {
 	persistentVolumeClaimCache    cache.Store
 	persistentVolumeClaimInformer cache.SharedIndexInformer
 
-	rsController *VMIReplicaSet
+	rsController *replicaset.VMIReplicaSet
 	rsInformer   cache.SharedIndexInformer
 
 	poolController *pool.PoolController
@@ -641,7 +642,7 @@ func (vca *VirtControllerApp) initCommon() {
 
 func (vca *VirtControllerApp) initReplicaSet() {
 	recorder := vca.newRecorder(k8sv1.NamespaceAll, "virtualmachinereplicaset-controller")
-	vca.rsController = NewVMIReplicaSet(vca.vmiInformer, vca.rsInformer, recorder, vca.clientSet, controller.BurstReplicas)
+	vca.rsController = replicaset.NewVMIReplicaSet(vca.vmiInformer, vca.rsInformer, recorder, vca.clientSet, controller.BurstReplicas)
 }
 
 func (vca *VirtControllerApp) initPool() {
