+++ pkg/virt-controller/watch/replicaset/replicaset.go
@@ -17,7 +17,7 @@
  *
  */
 
-package watch
+package replicaset
 
 import (
 	"context"
